Add prototype of the initial SIMD proposal at https://github.com/WebAssembly/simd/pull/1 It's enabled by default, but if you wish to disable it you can with the CMake ENABLE_SIMD_PROTOTYPE option. This also includes an implementation of the Blake2b hash function with the proposed SIMD operators in Test/Blake2b/blake2b.wast

+++ Source/WAST/NFA.h
@@ -71,7 +71,7 @@ namespace NFA
 		typedef int16 InternalStateIndex;
 		enum { internalMaxStates = INT16_MAX };
 
-		uint16 charToOffsetMap[256];
+		uint32 charToOffsetMap[256];
 		InternalStateIndex* stateAndOffsetToNextStateMap;
 		size_t numClasses;
 		size_t numStates;
